feat: 자판기가 보유하고 있는 금액 입력받는 기능 구현
