+++ src/main/java/vendingmachine/Application.java
@@ -1,7 +1,10 @@
 package vendingmachine;
 
+import vendingmachine.controller.VendingMachineController;
+
 public class Application {
-    public static void main(String[] args) {
-        // TODO: 프로그램 구현
-    }
+	public static void main(String[] args) {
+		VendingMachineController vendingMachineController = new VendingMachineController();
+		vendingMachineController.getVendingMachineAmount();
+	}
 }
