--- src/main/java/vendingmachine/Coin.java
@@ -1,16 +1,0 @@
-package vendingmachine;
-
-public enum Coin {
-    COIN_500(500),
-    COIN_100(100),
-    COIN_50(50),
-    COIN_10(10);
-
-    private final int amount;
-
-    Coin(final int amount) {
-        this.amount = amount;
-    }
-
-    // 추가 기능 구현
-}
