+++ src/main/java/vendingmachine/constant/InputMessage.java
@@ -1,0 +1,5 @@
+package vendingmachine.constant;
+
+public class InputMessage {
+	public static final String ASKING_VENDING_MACHINE_AMOUNT_MESSAGE = "자판기 있고 있는 금액을 입력해 주세요.";
+}
